task engine sets up dependencies and env variables in awsfirelens log router containers
