+++ agent/api/task/task_windows.go
@@ -16,7 +16,6 @@
 package task
 
 import (
-	"fmt"
 	"runtime"
 	"time"
 
@@ -27,12 +26,10 @@ import (
 	apicontainer "github.com/aws/amazon-ecs-agent/agent/api/container"
 	apicontainerstatus "github.com/aws/amazon-ecs-agent/agent/api/container/status"
 	apieni "github.com/aws/amazon-ecs-agent/agent/api/eni"
-	apierrors "github.com/aws/amazon-ecs-agent/agent/api/errors"
 	"github.com/aws/amazon-ecs-agent/agent/config"
 	"github.com/aws/amazon-ecs-agent/agent/credentials"
 	"github.com/aws/amazon-ecs-agent/agent/taskresource"
 	"github.com/aws/amazon-ecs-agent/agent/taskresource/credentialspec"
-	"github.com/aws/amazon-ecs-agent/agent/taskresource/firelens"
 	"github.com/aws/amazon-ecs-agent/agent/taskresource/fsxwindowsfileserver"
 	resourcestatus "github.com/aws/amazon-ecs-agent/agent/taskresource/status"
 	resourcetype "github.com/aws/amazon-ecs-agent/agent/taskresource/types"
@@ -51,13 +48,13 @@ const (
 	// firelensConfigBindFormatFluentd and firelensConfigBindFormatFluentbit specify the format of the firelens
 	// config file bind mount for fluentd and fluentbit firelens container respectively.
 	// First placeholder is host data dir, second placeholder is taskID.
-	firelensConfigBindFormatFluentd   = `%s\data\firelens\%s\config\fluent.conf:c:\data\fluentd\etc\fluent.conf`
-	firelensConfigBindFormatFluentbit = `%s\data\firelens\%s\config\fluent.conf:c:\data\fluent-bit\etc\fluent-bit.conf`
+	firelensConfigBindFormatFluentd   = `%s\firelens\%s\config\:C:\data\fluentd\etc\`
+	firelensConfigBindFormatFluentbit = `%s\firelens\%s\config\:C:\data\fluent-bit\etc\`
 
 	// firelensS3ConfigBindFormat specifies the format of the bind mount for the firelens config file downloaded from S3.
 	// First placeholder is host data dir, second placeholder is taskID, third placeholder is the s3 config path inside
 	// the firelens container.
-	firelensS3ConfigBindFormat = `%s\data\firelens\%s\config\external.conf:%s`
+	firelensS3ConfigBindFormat = `%s\firelens\%s\config\:%s`
 )
 
 // PlatformFields consists of fields specific to Windows for a task
@@ -311,33 +308,8 @@ func (task *Task) BuildCNIConfig(includeIPAMConfig bool, cniConfig *ecscni.Confi
 	return cniConfig, nil
 }
 
-// AddFirelensContainerBindMounts adds config file bind mount and socket directory bind mount to the firelens
-// container's host config.
-func (task *Task) AddFirelensContainerBindMounts(firelensConfig *apicontainer.FirelensConfig, hostConfig *dockercontainer.HostConfig,
-	config *config.Config) *apierrors.HostConfigError {
-	taskID, err := task.GetID()
-	if err != nil {
-		return &apierrors.HostConfigError{Msg: err.Error()}
-	}
-
-	var configBind, s3ConfigBind string
-	switch firelensConfig.Type {
-	case firelens.FirelensConfigTypeFluentd:
-		configBind = fmt.Sprintf(firelensConfigBindFormatFluentd, config.DataDirOnHost, taskID)
-		s3ConfigBind = fmt.Sprintf(firelensS3ConfigBindFormat, config.DataDirOnHost, taskID, firelens.S3ConfigPathFluentd)
-	case firelens.FirelensConfigTypeFluentbit:
-		configBind = fmt.Sprintf(firelensConfigBindFormatFluentbit, config.DataDirOnHost, taskID)
-		s3ConfigBind = fmt.Sprintf(firelensS3ConfigBindFormat, config.DataDirOnHost, taskID, firelens.S3ConfigPathFluentbit)
-	default:
-		return &apierrors.HostConfigError{Msg: fmt.Sprintf("encounter invalid firelens configuration type %s",
-			firelensConfig.Type)}
-	}
-
-	hostConfig.Binds = append(hostConfig.Binds, configBind)
-
-	// Add the s3 config bind mount if firelens container is using a config file from S3.
-	if firelensConfig.Options != nil && firelensConfig.Options[firelens.ExternalConfigTypeOption] == firelens.ExternalConfigTypeS3 {
-		hostConfig.Binds = append(hostConfig.Binds, s3ConfigBind)
-	}
-	return nil
+// AddFirelensSocketBindMount adds socket directory bind mount to the firelens container's host config.
+// sockets not supported on Windows
+func (task *Task) AddFirelensSocketAndS3BindMount(firelensConfig *apicontainer.FirelensConfig, hostConfig *dockercontainer.HostConfig,
+	config *config.Config, taskID string, s3ConfigBind string) {
 }
