+++ agent/engine/docker_task_engine.go
@@ -49,12 +49,10 @@ import (
 	"github.com/aws/amazon-ecs-agent/agent/statechange"
 	"github.com/aws/amazon-ecs-agent/agent/taskresource"
 	"github.com/aws/amazon-ecs-agent/agent/taskresource/credentialspec"
-	"github.com/aws/amazon-ecs-agent/agent/taskresource/firelens"
 	"github.com/aws/amazon-ecs-agent/agent/utils"
 	"github.com/aws/amazon-ecs-agent/agent/utils/retry"
 	utilsync "github.com/aws/amazon-ecs-agent/agent/utils/sync"
 	"github.com/aws/amazon-ecs-agent/agent/utils/ttime"
-	dockercontainer "github.com/docker/docker/api/types/container"
 
 	"github.com/cihub/seelog"
 	"github.com/docker/docker/api/types"
@@ -81,26 +79,6 @@ const (
 	maxEngineConnectRetryDelay         = 200 * time.Second
 	engineConnectRetryJitterMultiplier = 0.20
 	engineConnectRetryDelayMultiplier  = 1.5
-	// logDriverTypeFirelens is the log driver type for containers that want to use the firelens container to send logs.
-	logDriverTypeFirelens       = "awsfirelens"
-	logDriverTypeFluentd        = "fluentd"
-	logDriverTag                = "tag"
-	logDriverFluentdAddress     = "fluentd-address"
-	dataLogDriverPath           = "/data/firelens/"
-	logDriverAsyncConnect       = "fluentd-async-connect"
-	logDriverSubSecondPrecision = "fluentd-sub-second-precision"
-	logDriverBufferLimit        = "fluentd-buffer-limit"
-	dataLogDriverSocketPath     = "/socket/fluent.sock"
-	socketPathPrefix            = "unix://"
-
-	// fluentTagDockerFormat is the format for the log tag, which is "containerName-firelens-taskID"
-	fluentTagDockerFormat = "%s-firelens-%s"
-
-	// Environment variables are needed for firelens
-	fluentNetworkHost      = "FLUENT_HOST"
-	fluentNetworkPort      = "FLUENT_PORT"
-	FluentNetworkPortValue = "24224"
-	FluentAWSVPCHostValue  = "127.0.0.1"
 
 	defaultMonitorExecAgentsInterval = 15 * time.Minute
 
@@ -1081,50 +1059,9 @@ func (engine *DockerTaskEngine) createContainer(task *apitask.Task, container *a
 		}
 	}
 
-	firelensConfig := container.GetFirelensConfig()
-	if firelensConfig != nil {
-		err := task.AddFirelensContainerBindMounts(firelensConfig, hostConfig, engine.cfg)
-		if err != nil {
-			return dockerapi.DockerContainerMetadata{Error: apierrors.NamedError(err)}
-		}
-
-		cerr := task.PopulateSecretLogOptionsToFirelensContainer(container)
-		if cerr != nil {
-			return dockerapi.DockerContainerMetadata{Error: apierrors.NamedError(cerr)}
-		}
-
-		if firelensConfig.Type == firelens.FirelensConfigTypeFluentd {
-			// For fluentd router, needs to specify FLUENT_UID to root in order for the fluentd process to access
-			// the socket created by Docker.
-			container.MergeEnvironmentVariables(map[string]string{
-				"FLUENT_UID": "0",
-			})
-		}
-	}
-
-	// If the container is using a special log driver type "awsfirelens", it means the container wants to use
-	// the firelens container to send logs. In this case, override the log driver type to be fluentd
-	// and specify appropriate tag and fluentd-address, so that the logs are sent to and routed by the firelens container.
-	// Update the environment variables FLUENT_HOST and FLUENT_PORT depending on the supported network modes - bridge
-	// and awsvpc. For reference - https://docs.docker.com/config/containers/logging/fluentd/.
-	if hostConfig.LogConfig.Type == logDriverTypeFirelens {
-		hostConfig.LogConfig = getFirelensLogConfig(task, container, hostConfig, engine.cfg)
-		if task.IsNetworkModeAWSVPC() {
-			container.MergeEnvironmentVariables(map[string]string{
-				fluentNetworkHost: FluentAWSVPCHostValue,
-				fluentNetworkPort: FluentNetworkPortValue,
-			})
-		} else if container.GetNetworkModeFromHostConfig() == "" || container.GetNetworkModeFromHostConfig() == apitask.BridgeNetworkMode {
-			ipAddress, ok := getContainerHostIP(task.GetFirelensContainer().GetNetworkSettings())
-			if !ok {
-				err := apierrors.DockerClientConfigError{Msg: "unable to get BridgeIP for task in bridge  mode"}
-				return dockerapi.DockerContainerMetadata{Error: apierrors.NamedError(&err)}
-			}
-			container.MergeEnvironmentVariables(map[string]string{
-				fluentNetworkHost: ipAddress,
-				fluentNetworkPort: FluentNetworkPortValue,
-			})
-		}
+	firelensErr := engine.setupFireLensEnvironment(task, container, hostConfig)
+	if firelensErr.Error != nil {
+		return firelensErr
 	}
 
 	//Apply the log driver secret into container's LogConfig and Env secrets to container.Environment
@@ -1272,26 +1209,6 @@ func (engine *DockerTaskEngine) createContainer(task *apitask.Task, container *a
 		task.Arn, container.Name, metadata.DockerID, time.Since(createContainerBegin))
 	container.SetRuntimeID(metadata.DockerID)
 	return metadata
-}
-
-func getFirelensLogConfig(task *apitask.Task, container *apicontainer.Container, hostConfig *dockercontainer.HostConfig, cfg *config.Config) dockercontainer.LogConfig {
-	fields := strings.Split(task.Arn, "/")
-	taskID := fields[len(fields)-1]
-	tag := fmt.Sprintf(fluentTagDockerFormat, container.Name, taskID)
-	fluentd := socketPathPrefix + filepath.Join(cfg.DataDirOnHost, dataLogDriverPath, taskID, dataLogDriverSocketPath)
-	logConfig := hostConfig.LogConfig
-	bufferLimit, bufferLimitExists := logConfig.Config[apitask.FirelensLogDriverBufferLimitOption]
-	logConfig.Type = logDriverTypeFluentd
-	logConfig.Config = make(map[string]string)
-	logConfig.Config[logDriverTag] = tag
-	logConfig.Config[logDriverFluentdAddress] = fluentd
-	logConfig.Config[logDriverAsyncConnect] = strconv.FormatBool(true)
-	logConfig.Config[logDriverSubSecondPrecision] = strconv.FormatBool(true)
-	if bufferLimitExists {
-		logConfig.Config[logDriverBufferLimit] = bufferLimit
-	}
-	seelog.Debugf("Applying firelens log config for container %s: %v", container.Name, logConfig)
-	return logConfig
 }
 
 func (engine *DockerTaskEngine) startContainer(task *apitask.Task, container *apicontainer.Container) dockerapi.DockerContainerMetadata {
