+++ agent/engine/docker_task_engine_windows.go
@@ -16,11 +16,18 @@
 package engine
 
 import (
+	"fmt"
+	"strconv"
+	"strings"
 	"time"
 
 	apicontainer "github.com/aws/amazon-ecs-agent/agent/api/container"
+	apierrors "github.com/aws/amazon-ecs-agent/agent/api/errors"
 	apitask "github.com/aws/amazon-ecs-agent/agent/api/task"
+	"github.com/aws/amazon-ecs-agent/agent/config"
+	"github.com/aws/amazon-ecs-agent/agent/dockerclient/dockerapi"
 	"github.com/cihub/seelog"
+	dockercontainer "github.com/docker/docker/api/types/container"
 	"github.com/pkg/errors"
 )
 
@@ -28,6 +35,25 @@ const (
 	// Constants for CNI timeout during setup and cleanup on Windows.
 	cniSetupTimeout   = 3 * time.Minute
 	cniCleanupTimeout = 2 * time.Minute
+
+	// logDriverTypeFirelens is the log driver type for containers that want to use the firelens container to send logs.
+	logDriverTypeFirelens       = "awsfirelens"
+	logDriverTypeFluentd        = "fluentd"
+	logDriverTag                = "tag"
+	logDriverFluentdAddress     = "fluentd-address"
+	dataLogDriverPath           = "/data/firelens/"
+	logDriverAsyncConnect       = "fluentd-async-connect"
+	logDriverSubSecondPrecision = "fluentd-sub-second-precision"
+	logDriverBufferLimit        = "fluentd-buffer-limit"
+
+	// fluentTagDockerFormat is the format for the log tag, which is "containerName-firelens-taskID"
+	fluentTagDockerFormat = "%s-firelens-%s"
+	hostPortFormat        = "%s:%s"
+
+	// Environment variables are needed for firelens
+	fluentNetworkHost      = "FLUENT_HOST"
+	fluentNetworkPort      = "FLUENT_PORT"
+	FluentNetworkPortValue = "24224"
 )
 
 func (engine *DockerTaskEngine) updateTaskENIDependencies(task *apitask.Task) {
@@ -57,4 +83,104 @@ func (engine *DockerTaskEngine) invokePluginsForContainer(task *apitask.Task, co
 	}
 
 	return nil
+}
+
+func (engine *DockerTaskEngine) setupFireLensEnvironment(task *apitask.Task, container *apicontainer.Container, hostConfig *dockercontainer.HostConfig) dockerapi.DockerContainerMetadata {
+	firelensConfig := container.GetFirelensConfig()
+	if firelensConfig != nil {
+		err := task.AddFirelensContainerBindMounts(firelensConfig, hostConfig, engine.cfg)
+		if err != nil {
+			return dockerapi.DockerContainerMetadata{Error: apierrors.NamedError(err)}
+		}
+
+		cerr := task.PopulateSecretLogOptionsToFirelensContainer(container)
+		if cerr != nil {
+			return dockerapi.DockerContainerMetadata{Error: apierrors.NamedError(cerr)}
+		}
+	}
+
+	// If the container is using a special log driver type "awsfirelens", it means the container wants to use
+	// the firelens container to send logs. In this case, override the log driver type to be fluentd
+	// and specify appropriate tag and fluentd-address, so that the logs are sent to and routed by the firelens container.
+	// Update the environment variables FLUENT_HOST and FLUENT_PORT depending on the supported network modes - bridge
+	// and awsvpc. For reference - https://docs.docker.com/config/containers/logging/fluentd/.
+	if hostConfig.LogConfig.Type == logDriverTypeFirelens {
+		hostConfig.LogConfig = getFirelensLogConfig(task, container, hostConfig, engine.cfg)
+		if task.IsNetworkModeAWSVPC() {
+			primaryENI := task.GetPrimaryENI()
+			if primaryENI == nil {
+				err := apierrors.DockerClientConfigError{Msg: "unable to get ENI for task in AWSVPC mode"}
+				return dockerapi.DockerContainerMetadata{Error: apierrors.NamedError(&err)}
+			}
+			fluentIP := primaryENI.GetPrimaryIPv4Address()
+			if fluentIP == "" {
+				err := apierrors.DockerClientConfigError{Msg: "ENI does not have a primary IPv4 address"}
+				return dockerapi.DockerContainerMetadata{Error: apierrors.NamedError(&err)}
+			}
+			container.MergeEnvironmentVariables(map[string]string{
+				fluentNetworkHost: fluentIP,
+				fluentNetworkPort: FluentNetworkPortValue,
+			})
+		} else if container.GetNetworkModeFromHostConfig() == "" || container.GetNetworkModeFromHostConfig() == apitask.BridgeNetworkMode {
+			ipAddress := ""
+			for _, network := range task.GetFirelensContainer().GetNetworkSettings().Networks {
+				if network.IPAddress != "" {
+					ipAddress = network.IPAddress
+				}
+			}
+			if ipAddress == "" {
+				ipAddress = task.GetFirelensContainer().GetNetworkSettings().DefaultNetworkSettings.IPAddress
+			}
+			if ipAddress == "" {
+				err := apierrors.DockerClientConfigError{Msg: "unable to get BridgeIP for task in bridge  mode"}
+				return dockerapi.DockerContainerMetadata{Error: apierrors.NamedError(&err)}
+			}
+			container.MergeEnvironmentVariables(map[string]string{
+				fluentNetworkHost: ipAddress,
+				fluentNetworkPort: FluentNetworkPortValue,
+			})
+		}
+	}
+	return dockerapi.DockerContainerMetadata{}
+}
+
+func getFirelensLogConfig(task *apitask.Task, container *apicontainer.Container, hostConfig *dockercontainer.HostConfig, cfg *config.Config) dockercontainer.LogConfig {
+	fluentAddr := ""
+	// ok := false
+	if task.IsNetworkModeAWSVPC() {
+		primaryENI := task.GetPrimaryENI()
+		if primaryENI == nil {
+			seelog.Error("unable to get ENI for task in AWSVPC mode")
+		}
+		fluentAddr = task.GetPrimaryENI().GetPrimaryIPv4Address()
+	} else if container.GetNetworkModeFromHostConfig() == "" || container.GetNetworkModeFromHostConfig() == apitask.BridgeNetworkMode {
+		// fluentAddr, ok = getContainerHostIP(task.GetFirelensContainer().GetNetworkSettings())
+		for _, network := range task.GetFirelensContainer().GetNetworkSettings().Networks {
+			if network.IPAddress != "" {
+				fluentAddr = network.IPAddress
+			}
+		}
+		if fluentAddr == "" {
+			fluentAddr = task.GetFirelensContainer().GetNetworkSettings().DefaultNetworkSettings.IPAddress
+		}
+		if fluentAddr == "" {
+			seelog.Error("unable to get BridgeIP of log router")
+		}
+	}
+	fields := strings.Split(task.Arn, "/")
+	taskID := fields[len(fields)-1]
+	tag := fmt.Sprintf(fluentTagDockerFormat, container.Name, taskID)
+	logConfig := hostConfig.LogConfig
+	bufferLimit, bufferLimitExists := logConfig.Config[apitask.FirelensLogDriverBufferLimitOption]
+	logConfig.Type = logDriverTypeFluentd
+	logConfig.Config = make(map[string]string)
+	logConfig.Config[logDriverTag] = tag
+	logConfig.Config[logDriverFluentdAddress] = fmt.Sprintf(hostPortFormat, fluentAddr, FluentNetworkPortValue)
+	logConfig.Config[logDriverAsyncConnect] = strconv.FormatBool(true)
+	logConfig.Config[logDriverSubSecondPrecision] = strconv.FormatBool(true)
+	if bufferLimitExists {
+		logConfig.Config[logDriverBufferLimit] = bufferLimit
+	}
+	seelog.Debugf("Applying firelens log config for container %s: %v", container.Name, logConfig)
+	return logConfig
 }
