+++ agent/taskresource/firelens/firelens_unimplemented.go
@@ -15,25 +15,15 @@
 
 package firelens
 
-import generator "github.com/awslabs/go-config-generator-for-fluentd-and-fluentbit"
+import (
+	"github.com/aws/amazon-ecs-agent/agent/utils/oswrapper"
+)
 
-// generateConfig generates a FluentConfig object that contains all necessary information to construct
-// a fluentd or fluentbit config file for a firelens container.
-func (firelens *FirelensResource) generateConfig() (generator.FluentConfig, error) {
-	config := generator.New()
-	return config, nil
+func (firelens *FirelensResource) createSocketDirectories() error {
+	return nil
 }
 
-// createDirectories creates two directories:
-//  - $(DATA_DIR)/firelens/$(TASK_ID)/config: used to store firelens config file. The config file under this directory
-//    will be mounted to the firelens container at an expected path.
-//  - $(DATA_DIR)/firelens/$(TASK_ID)/socket: used to store the unix socket. This directory will be mounted to
-//    the firelens container and it will generate a socket file under this directory. Containers that use firelens to
-//    send logs will then use this socket to send logs to the firelens container.
-// Note: socket path has a limit of at most 108 characters on Linux. If using default data dir, the
-// resulting socket path will be 79 characters (/var/lib/ecs/data/firelens/<task-id>/socket/fluent.sock) which is fine.
-// However if ECS_HOST_DATA_DIR is specified to be a longer path, we will exceed the limit and fail. I don't really
-// see a way to avoid this failure since ECS_HOST_DATA_DIR can be arbitrary long..
-func (firelens *FirelensResource) createDirectories() error {
+// writeConfigFile writes a config file at a given path.
+func (firelens *FirelensResource) writeConfigFile(writeFunc func(file oswrapper.File) error, filePath string) error {
 	return nil
 }
