+++ agent/taskresource/firelens/firelens_unix.go
@@ -18,32 +18,48 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/pkg/errors"
+	"github.com/aws/amazon-ecs-agent/agent/utils/oswrapper"
 )
 
-var mkdirAll = os.MkdirAll
-
-// createDirectories creates two directories:
-//  - $(DATA_DIR)/firelens/$(TASK_ID)/config: used to store firelens config file. The config file under this directory
-//    will be mounted to the firelens container at an expected path.
-//  - $(DATA_DIR)/firelens/$(TASK_ID)/socket: used to store the unix socket. This directory will be mounted to
-//    the firelens container and it will generate a socket file under this directory. Containers that use firelens to
-//    send logs will then use this socket to send logs to the firelens container.
-// Note: socket path has a limit of at most 108 characters on Linux. If using default data dir, the
-// resulting socket path will be 79 characters (/var/lib/ecs/data/firelens/<task-id>/socket/fluent.sock) which is fine.
-// However if ECS_HOST_DATA_DIR is specified to be a longer path, we will exceed the limit and fail. I don't really
-// see a way to avoid this failure since ECS_HOST_DATA_DIR can be arbitrary long..
-func (firelens *FirelensResource) createDirectories() error {
-	configDir := filepath.Join(firelens.resourceDir, "config")
-	err := mkdirAll(configDir, os.ModePerm)
-	if err != nil {
-		return errors.Wrap(err, "unable to create config directory")
-	}
-
+func (firelens *FirelensResource) createSocketDirectories() error {
 	socketDir := filepath.Join(firelens.resourceDir, "socket")
-	err = mkdirAll(socketDir, os.ModePerm)
+	err := mkdirAll(socketDir, os.ModePerm)
 	if err != nil {
-		return errors.Wrap(err, "unable to create socket directory")
+		return err
 	}
+	return nil
+}
+
+var rename = os.Rename
+
+// writeConfigFile writes a config file at a given path.
+func (firelens *FirelensResource) writeConfigFile(writeFunc func(file oswrapper.File) error, filePath string) error {
+	temp, err := firelens.ioutil.TempFile(firelens.resourceDir, tempFile)
+	if err != nil {
+		return err
+	}
+	defer temp.Close()
+
+	err = writeFunc(temp)
+	if err != nil {
+		return err
+	}
+
+	err = temp.Chmod(os.FileMode(configFilePerm))
+	if err != nil {
+		return err
+	}
+
+	// Persist the config file to disk.
+	err = temp.Sync()
+	if err != nil {
+		return err
+	}
+
+	err = rename(temp.Name(), filePath)
+	if err != nil {
+		return err
+	}
+
 	return nil
 }
