+++ agent/taskresource/firelens/firelensconfig.go
@@ -16,6 +16,7 @@ package firelens
 import (
 	"fmt"
 
+	"github.com/cihub/seelog"
 	"github.com/pkg/errors"
 
 	generator "github.com/awslabs/go-config-generator-for-fluentd-and-fluentbit"
@@ -28,6 +29,9 @@ const (
 	// FirelensConfigTypeFluentbit is the type of a fluentbit firelens container.
 	FirelensConfigTypeFluentbit = "fluentbit"
 
+	// inputNameForward is the name of the tcp socket input plugin for fluentd and fluentbit.
+	inputNameForward = "forward"
+
 	// outputTypeLogOptionKeyFluentd is the key for the log option that specifies output plugin type for fluentd.
 	outputTypeLogOptionKeyFluentd = "@type"
 
@@ -39,7 +43,99 @@ const (
 
 	// excludePatternKey is the key for exclude pattern.
 	excludePatternKey = "exclude-pattern"
+
+	// fluentTagOutputFormat is the format for the log tag captured by the output section. First placeholder is
+	// container name. Second placeholder is the wildcard that matches all contents.
+	// When customer uses config generated by the agent, the input log will have tag as containerName-firelens-taskID which
+	// matches this. Customers can also generate logs with their own config and specify the tag to start with containerName-firelens
+	// to match this.
+	fluentTagOutputFormat = "%s-firelens%s"
+
+	// matchAnyWildcardFluentd is the wildcard to match arbitrary content for fluentd.
+	matchAnyWildcardFluentd = "**"
+	// matchAnyWildcardFluentd is the wildcard to match arbitrary content for fluentbit.
+	matchAnyWildcardFluentbit = "*"
+
+	// inputBindOptionFluentd is the key for specifying host for fluentd for tcp.
+	inputBindOptionFluentd = "bind"
+
+	// inputPortOptionFluentd is the key for specifying port for fluentd for tcp.
+	inputPortOptionFluentd = "port"
+
+	// inputPortValue is the value for specifying port for fluentd for tcp.
+	inputPortValue = "24224"
+
+	// inputListenOptionFluentbit is the key for the log option that specifies host for fluentbit.
+	inputListenOptionFluentbit = "Listen"
+
+	// inputPortOptionFluentbit is the key for the log option that specifies port for fluentbit.
+	inputPortOptionFluentbit = "Port"
+
+	// bridgeNetworkMode specifies bridge type mode for a task
+	bridgeNetworkMode = "bridge"
+
+	// specifies awsvpc type mode for a task
+	awsvpcNetworkMode = "awsvpc"
+
+	// healthcheckInputNameFluentbit in the input plugin used to receive health check message for fluentbit.
+	healthcheckInputNameFluentbit = "tcp"
+	// healthcheckInputBindValue is the source where health check message comes from.
+	healthcheckInputBindValue = "127.0.0.1"
+	// healthcheckInputPortValue is the port for healthcheck.
+	healthcheckInputPortValue = "8877"
+	// healthcheckTag is the tag for health check message.
+	healthcheckTag = "firelens-healthcheck"
+	// healthcheckOutputName is the output plugin that health check message goes to. It's a black hole so that the health
+	// check messages don't go into logs.
+	healthcheckOutputName = "null"
 )
+
+// generateConfig generates a FluentConfig object that contains all necessary information to construct
+// a fluentd or fluentbit config file for a firelens container.
+func (firelens *FirelensResource) generateConfig() (generator.FluentConfig, error) {
+	config := generator.New()
+
+	var matchAnyWildcard string
+	if firelens.firelensConfigType == FirelensConfigTypeFluentd {
+		matchAnyWildcard = matchAnyWildcardFluentd
+	} else {
+		matchAnyWildcard = matchAnyWildcardFluentbit
+	}
+
+	firelens.addSocketInput(config)
+	firelens.addTCPInput(config)
+
+	if firelens.ecsMetadataEnabled {
+		// Add ecs metadata fields to the log stream.
+		config.AddFieldToRecord("ecs_cluster", firelens.cluster, matchAnyWildcard).
+			AddFieldToRecord("ecs_task_arn", firelens.taskARN, matchAnyWildcard).
+			AddFieldToRecord("ecs_task_definition", firelens.taskDefinition, matchAnyWildcard)
+		if firelens.ec2InstanceID != "" {
+			config.AddFieldToRecord("ec2_instance_id", firelens.ec2InstanceID, matchAnyWildcard)
+		}
+	}
+
+	// Specify log stream output. Each container that uses the firelens container to stream logs
+	// may have its own output section with options, constructed from container's log options.
+	for containerName, logOptions := range firelens.containerToLogOptions {
+		tag := fmt.Sprintf(fluentTagOutputFormat, containerName, matchAnyWildcard) // Each output section is distinguished by a tag specific to a container.
+		newConfig, err := addOutputSection(tag, firelens.firelensConfigType, logOptions, config)
+		if err != nil {
+			return nil, fmt.Errorf("unable to apply log options of container %s to firelens config: %v", containerName, err)
+		}
+		config = newConfig
+	}
+
+	// Include external config file if specified.
+	if firelens.externalConfigType == ExternalConfigTypeFile {
+		config.AddExternalConfig(firelens.externalConfigValue, generator.AfterFilters)
+	} else if firelens.externalConfigType == ExternalConfigTypeS3 {
+		config.AddExternalConfig(firelens.getS3ConfPath(), generator.AfterFilters)
+	}
+	seelog.Infof("Included external firelens config file at: %s", firelens.externalConfigValue)
+
+	return config, nil
+}
 
 // addOutputSection adds an output section to the firelens container's config that specifies how it routes another
 // container's logs. It's constructed based on that container's log options.
@@ -85,4 +181,34 @@ func addOutputSection(tag, firelensConfigType string, logOptions map[string]stri
 	// Output key is specified. Add an output section.
 	config.AddOutput(output, tag, outputOptions)
 	return config, nil
+}
+
+// Specify log stream input of tcp socket kind that can be used for communication between the Firelens
+// container and other containers if the network is bridge or awsvpc mode. Also add health check sections to support
+// doing container health check on firlens container for these two modes.
+func (firelens *FirelensResource) addTCPInput(config generator.FluentConfig) {
+	if firelens.networkMode == bridgeNetworkMode || firelens.networkMode == awsvpcNetworkMode || firelens.networkMode == "" {
+		var inputMap map[string]string
+		var inputName, inputBindValue string
+		if firelens.networkMode == bridgeNetworkMode || firelens.networkMode == "" {
+			inputBindValue = inputBridgeBindValue
+		} else if firelens.networkMode == awsvpcNetworkMode {
+			inputBindValue = inputAWSVPCBindValue
+		}
+		if firelens.firelensConfigType == FirelensConfigTypeFluentd {
+			inputMap = map[string]string{
+				inputPortOptionFluentd: inputPortValue,
+				inputBindOptionFluentd: inputBindValue,
+			}
+			inputName = inputNameForward
+		} else {
+			inputName = inputNameForward
+			inputMap = map[string]string{
+				inputPortOptionFluentbit:   inputPortValue,
+				inputListenOptionFluentbit: inputBindValue,
+			}
+		}
+		config.AddInput(inputName, "", inputMap)
+		firelens.addHealthcheckSections(config)
+	}
 }
