+++ agent/taskresource/firelens/firelensconfig_unix.go
@@ -15,18 +15,12 @@
 package firelens
 
 import (
-	"fmt"
-
 	generator "github.com/awslabs/go-config-generator-for-fluentd-and-fluentbit"
-	"github.com/cihub/seelog"
 )
 
 const (
 	// socketInputNameFluentd is the name of the socket input plugin for fluentd.
 	socketInputNameFluentd = "unix"
-
-	// inputNameForward is the name of the tcp socket input plugin for fluentd and fluentbit.
-	inputNameForward = "forward"
 
 	// socketInputPathOptionFluentd is the key for specifying socket path for fluentd.
 	socketInputPathOptionFluentd = "path"
@@ -41,143 +35,27 @@ const (
 	S3ConfigPathFluentd   = "/fluentd/etc/external.conf"
 	S3ConfigPathFluentbit = "/fluent-bit/etc/external.conf"
 
-	// fluentTagOutputFormat is the format for the log tag captured by the output section. First placeholder is
-	// container name. Second placeholder is the wildcard that matches all contents.
-	// When customer uses config generated by the agent, the input log will have tag as containerName-firelens-taskID which
-	// matches this. Customers can also generate logs with their own config and specify the tag to start with containerName-firelens
-	// to match this.
-	fluentTagOutputFormat = "%s-firelens%s"
-
-	// matchAnyWildcardFluentd is the wildcard to match arbitrary content for fluentd.
-	matchAnyWildcardFluentd = "**"
-	// matchAnyWildcardFluentd is the wildcard to match arbitrary content for fluentbit.
-	matchAnyWildcardFluentbit = "*"
-
-	// inputBindOptionFluentd is the key for specifying host for fluentd for tcp.
-	inputBindOptionFluentd = "bind"
-
 	// inputBridgeBindValue is the value for specifying host for Bridge mode.
 	inputBridgeBindValue = "0.0.0.0"
 
 	// inputAWSVPCBindValue is the value for specifying host for AWSVPC mode.
 	inputAWSVPCBindValue = "127.0.0.1"
-
-	// inputPortOptionFluentd is the key for specifying port for fluentd for tcp.
-	inputPortOptionFluentd = "port"
-
-	// inputPortValue is the value for specifying port for fluentd for tcp.
-	inputPortValue = "24224"
-
-	// healthcheckInputNameFluentbit in the input plugin used to receive health check message for fluentbit.
-	healthcheckInputNameFluentbit = "tcp"
-	// healthcheckInputBindValue is the source where health check message comes from.
-	healthcheckInputBindValue = "127.0.0.1"
-	// healthcheckInputPortValue is the port for healthcheck.
-	healthcheckInputPortValue = "8877"
-	// healthcheckTag is the tag for health check message.
-	healthcheckTag = "firelens-healthcheck"
-	// healthcheckOutputName is the output plugin that health check message goes to. It's a black hole so that the health
-	// check messages don't go into logs.
-	healthcheckOutputName = "null"
-
-	// inputListenOptionFluentbit is the key for the log option that specifies host for fluentbit.
-	inputListenOptionFluentbit = "Listen"
-
-	// inputPortOptionFluentbit is the key for the log option that specifies port for fluentbit.
-	inputPortOptionFluentbit = "Port"
-
-	// bridgeNetworkMode specifies bridge type mode for a task
-	bridgeNetworkMode = "bridge"
-
-	// specifies awsvpc type mode for a task
-	awsvpcNetworkMode = "awsvpc"
 )
 
-// generateConfig generates a FluentConfig object that contains all necessary information to construct
-// a fluentd or fluentbit config file for a firelens container.
-func (firelens *FirelensResource) generateConfig() (generator.FluentConfig, error) {
-	config := generator.New()
-
-	// Specify log stream input, which is a unix socket that will be used for communication between the Firelens
-	// container and other containers.
-	var inputName, inputPathOption, matchAnyWildcard string
+// Specify log stream input, which is a unix socket that will be used for communication between the Firelens
+// container and other containers.
+func (firelens *FirelensResource) addSocketInput(config generator.FluentConfig) {
+	var inputName, inputPathOption string
 	if firelens.firelensConfigType == FirelensConfigTypeFluentd {
 		inputName = socketInputNameFluentd
 		inputPathOption = socketInputPathOptionFluentd
-		matchAnyWildcard = matchAnyWildcardFluentd
 	} else {
 		inputName = inputNameForward
 		inputPathOption = socketInputPathOptionFluentbit
-		matchAnyWildcard = matchAnyWildcardFluentbit
 	}
 	config.AddInput(inputName, "", map[string]string{
 		inputPathOption: socketPath,
 	})
-	// Specify log stream input of tcp socket kind that can be used for communication between the Firelens
-	// container and other containers if the network is bridge or awsvpc mode. Also add health check sections to support
-	// doing container health check on firlens container for these two modes.
-	if firelens.networkMode == bridgeNetworkMode || firelens.networkMode == awsvpcNetworkMode {
-		var inputMap map[string]string
-		var inputBindValue string
-		if firelens.networkMode == bridgeNetworkMode {
-			inputBindValue = inputBridgeBindValue
-		} else if firelens.networkMode == awsvpcNetworkMode {
-			inputBindValue = inputAWSVPCBindValue
-		}
-		if firelens.firelensConfigType == FirelensConfigTypeFluentd {
-			inputMap = map[string]string{
-				inputPortOptionFluentd: inputPortValue,
-				inputBindOptionFluentd: inputBindValue,
-			}
-			inputName = inputNameForward
-		} else {
-			inputName = inputNameForward
-			inputMap = map[string]string{
-				inputPortOptionFluentbit:   inputPortValue,
-				inputListenOptionFluentbit: inputBindValue,
-			}
-		}
-		config.AddInput(inputName, "", inputMap)
-
-		firelens.addHealthcheckSections(config)
-	}
-
-	if firelens.ecsMetadataEnabled {
-		// Add ecs metadata fields to the log stream.
-		config.AddFieldToRecord("ecs_cluster", firelens.cluster, matchAnyWildcard).
-			AddFieldToRecord("ecs_task_arn", firelens.taskARN, matchAnyWildcard).
-			AddFieldToRecord("ecs_task_definition", firelens.taskDefinition, matchAnyWildcard)
-		if firelens.ec2InstanceID != "" {
-			config.AddFieldToRecord("ec2_instance_id", firelens.ec2InstanceID, matchAnyWildcard)
-		}
-	}
-
-	// Specify log stream output. Each container that uses the firelens container to stream logs
-	// may have its own output section with options, constructed from container's log options.
-	for containerName, logOptions := range firelens.containerToLogOptions {
-		tag := fmt.Sprintf(fluentTagOutputFormat, containerName, matchAnyWildcard) // Each output section is distinguished by a tag specific to a container.
-		newConfig, err := addOutputSection(tag, firelens.firelensConfigType, logOptions, config)
-		if err != nil {
-			return nil, fmt.Errorf("unable to apply log options of container %s to firelens config: %v", containerName, err)
-		}
-		config = newConfig
-	}
-
-	// Include external config file if specified.
-	if firelens.externalConfigType == ExternalConfigTypeFile {
-		config.AddExternalConfig(firelens.externalConfigValue, generator.AfterFilters)
-	} else if firelens.externalConfigType == ExternalConfigTypeS3 {
-		var s3ConfPath string
-		if firelens.firelensConfigType == FirelensConfigTypeFluentd {
-			s3ConfPath = S3ConfigPathFluentd
-		} else {
-			s3ConfPath = S3ConfigPathFluentbit
-		}
-		config.AddExternalConfig(s3ConfPath, generator.AfterFilters)
-	}
-	seelog.Infof("Included external firelens config file at: %s", firelens.externalConfigValue)
-
-	return config, nil
 }
 
 // addHealthcheckSections adds a health check input section and a health check output section to the config.
@@ -197,4 +75,12 @@ func (firelens *FirelensResource) addHealthcheckSections(config generator.Fluent
 
 	// Add healthcheck output section.
 	config.AddOutput(healthcheckOutputName, healthcheckTag, nil)
+}
+
+func (firelens *FirelensResource) getS3ConfPath() string {
+	if firelens.firelensConfigType == FirelensConfigTypeFluentd {
+		return S3ConfigPathFluentd
+	} else {
+		return S3ConfigPathFluentbit
+	}
 }
